Add a pre-commit hook to format texts consistently
The hook embeds a Python script that will sort attributes by name
and indent each level with one space.  Previously we used a
stylesheet to sort the attributes and xmllint to do the indent,
but this meant two passes through the file and xmllint would fall
down badly in some cases and string together pages of text as a
single line.