Combine some texts and restore missing poetry
Quite a few collections of mostly plays had the poetry parts
excised at some point and this restores some of them.

There are also a lot of rewritten subrepo commits that were already
there in the previous commit because git doesn't know that amending
something that hasn't changed isn't an actual change.